refactor(runner): split network log manager ownership
